fix: fixes for login in backend 🐛
