+++ server/middleware/passport/errors/USER_ACCOUNT_DISABLED.ts
@@ -1,0 +1,12 @@
+import { SigninError } from './SigninError'
+
+/**
+ * User account disabled error
+ */
+
+export const USER_ACCOUNT_DISABLED = new SigninError(
+  'e0666582',
+  'An error occured signing you in',
+  'Your account is disabled.',
+  'Cancel'
+)
