+++ server/middleware/passport/microsoft/processUserInvitation.ts
@@ -3,10 +3,12 @@ import { ExternalUserInvitationInput, User } from 'server/graphql'
 import { SubscriptionService, UserService } from 'server/services'
 import { debug, PROVIDER } from './onVerifySignin'
 import { IProfileJson } from './types'
+import _ from 'lodash'
+import { TENANT_NOT_ENROLLED } from '../errors'
 
 /**
  * Process a user invitation and create a user account
- * 
+ *
  * @param userSrv - User service instance
  * @param subSvc - Subscription service instance
  * @param profile - User profile object
@@ -22,6 +24,10 @@ export async function processUserInvitation(
     `User invitation with ID ${userInvitation.id} found. Adding user to the database`
   )
 
+  throw TENANT_NOT_ENROLLED
+
+  const invitedBy = await userSrv.getById(userInvitation.invitedBy)
+
   const dbUser: Partial<User> = {
     id: profile.oid,
     mail: profile.preferred_username,
@@ -30,6 +36,8 @@ export async function processUserInvitation(
     preferredLanguage: userInvitation.preferredLanguage,
     tenantId: profile.tid,
     isExternal: true,
+    invitationId: userInvitation.id,
+    invitedAt: userInvitation.invitedAt,
     startPage: userInvitation.startPage,
     configuration: {
       ui: {
@@ -38,10 +46,20 @@ export async function processUserInvitation(
     }
   }
 
+  if (Boolean(invitedBy)) {
+    dbUser.manager = _.pick(invitedBy, ['id', 'mail', 'displayName'])
+  }
+
   await userSrv.addUser(dbUser)
 
-  debug(`Registering user ${profile.preferred_username} (${profile.oid}) with subscription ${userInvitation.subscription.id}`)
-  await subSvc.registerExternalUser(PROVIDER, profile.oid, userInvitation.subscription.id)
+  debug(
+    `Registering user ${profile.preferred_username} (${profile.oid}) with subscription ${userInvitation.subscription.id}`
+  )
+  await subSvc.registerExternalUser(
+    PROVIDER,
+    profile.oid,
+    userInvitation.subscription.id
+  )
 
   debug(`Removing external invitation with ID ${userInvitation.id}`)
   await subSvc.removeExternalInvitation(userInvitation)
