+++ server/middleware/passport/microsoft/types.ts
@@ -1,6 +1,6 @@
 export interface IProfileJson {
-    preferred_username: string
-    oid: string
-    displayName: string
-    tid: string
+  preferred_username: string
+  oid: string
+  displayName: string
+  tid: string
 }
